+++ src/ui-components/SubjectsCreateForm.jsx
@@ -542,9 +542,9 @@ export default function SubjectsCreateForm(props) {
         step="any"
         value={current_Grade}
         onChange={(e) => {
-          let value = isNaN(parseInt(e.target.value))
+          let value = isNaN(parseFloat(e.target.value))
             ? e.target.value
-            : parseInt(e.target.value);
+            : parseFloat(e.target.value);
           if (onChange) {
             const modelFields = {
               subject_Name,
@@ -575,9 +575,9 @@ export default function SubjectsCreateForm(props) {
         step="any"
         value={target_Grade}
         onChange={(e) => {
-          let value = isNaN(parseInt(e.target.value))
+          let value = isNaN(parseFloat(e.target.value))
             ? e.target.value
-            : parseInt(e.target.value);
+            : parseFloat(e.target.value);
           if (onChange) {
             const modelFields = {
               subject_Name,
