+++ src/ui-components/TaskGradeInfoCreateForm.jsx
@@ -425,9 +425,9 @@ export default function TaskGradeInfoCreateForm(props) {
         step="any"
         value={current_Grade}
         onChange={(e) => {
-          let value = isNaN(parseInt(e.target.value))
+          let value = isNaN(parseFloat(e.target.value))
             ? e.target.value
-            : parseInt(e.target.value);
+            : parseFloat(e.target.value);
           if (onChange) {
             const modelFields = {
               current_Grade: value,
@@ -458,9 +458,9 @@ export default function TaskGradeInfoCreateForm(props) {
         step="any"
         value={task_Weightage}
         onChange={(e) => {
-          let value = isNaN(parseInt(e.target.value))
+          let value = isNaN(parseFloat(e.target.value))
             ? e.target.value
-            : parseInt(e.target.value);
+            : parseFloat(e.target.value);
           if (onChange) {
             const modelFields = {
               current_Grade,
@@ -491,9 +491,9 @@ export default function TaskGradeInfoCreateForm(props) {
         step="any"
         value={overall_Percentage}
         onChange={(e) => {
-          let value = isNaN(parseInt(e.target.value))
+          let value = isNaN(parseFloat(e.target.value))
             ? e.target.value
-            : parseInt(e.target.value);
+            : parseFloat(e.target.value);
           if (onChange) {
             const modelFields = {
               current_Grade,
